MAINT: move elementwise algoirthm implementations to appropriate files
