+++ scipy/integrate/tests/meson.build
@@ -6,7 +6,8 @@ py3.install_sources([
     'test_integrate.py',
     'test_odeint_jac.py',
     'test_quadpack.py',
-    'test_quadrature.py'
+    'test_quadrature.py',
+    'test_tanhsinh.py'
   ],
   subdir: 'scipy/integrate/tests'
 )
