+++ scipy/optimize/_zeros_py.py
@@ -4,8 +4,6 @@ import operator
 from . import _zeros
 from ._optimize import OptimizeResult
 import numpy as np
-from scipy._lib._elementwise_algorithm import (_scalar_optimization_initialize,
-                                               _scalar_optimization_loop)
 
 
 _iter = 100
@@ -18,13 +16,10 @@ __all__ = ['newton', 'bisect', 'ridder', 'brentq', 'brenth', 'toms748',
 # Must agree with CONVERGED, SIGNERR, CONVERR, ...  in zeros.h
 _ECONVERGED = 0
 _ESIGNERR = -1  # used in _chandrupatla
-_EERRORINCREASE = -1  # used in _differentiate
-_ELIMITS = -1  # used in _bracket_root
 _ECONVERR = -2
 _EVALUEERR = -3
 _ECALLBACK = -4
 _EINPROGRESS = 1
-_ESTOPONESIDE = 2  # used in _bracket_root
 
 CONVERGED = 'converged'
 SIGNERR = 'sign error'
@@ -1406,1202 +1401,3 @@ def toms748(f, a, b, args=(), k=1,
     x, function_calls, iterations, flag = result
     return _results_select(full_output, (x, function_calls, iterations, flag),
                            "toms748")
-
-
-def _bracket_root_iv(func, a, b, min, max, factor, args, maxiter):
-
-    if not callable(func):
-        raise ValueError('`func` must be callable.')
-
-    if not np.iterable(args):
-        args = (args,)
-
-    a = np.asarray(a)[()]
-    if not np.issubdtype(a.dtype, np.number) or np.iscomplex(a).any():
-        raise ValueError('`a` must be numeric and real.')
-
-    b = a + 1 if b is None else b
-    min = -np.inf if min is None else min
-    max = np.inf if max is None else max
-    factor = 2. if factor is None else factor
-    a, b, min, max, factor = np.broadcast_arrays(a, b, min, max, factor)
-
-    if not np.issubdtype(b.dtype, np.number) or np.iscomplex(b).any():
-        raise ValueError('`b` must be numeric and real.')
-
-    if not np.issubdtype(min.dtype, np.number) or np.iscomplex(min).any():
-        raise ValueError('`min` must be numeric and real.')
-
-    if not np.issubdtype(max.dtype, np.number) or np.iscomplex(max).any():
-        raise ValueError('`max` must be numeric and real.')
-
-    if not np.issubdtype(factor.dtype, np.number) or np.iscomplex(factor).any():
-        raise ValueError('`factor` must be numeric and real.')
-    if not np.all(factor > 1):
-        raise ValueError('All elements of `factor` must be greater than 1.')
-
-    maxiter = np.asarray(maxiter)
-    message = '`maxiter` must be a non-negative integer.'
-    if (not np.issubdtype(maxiter.dtype, np.number) or maxiter.shape != tuple()
-            or np.iscomplex(maxiter)):
-        raise ValueError(message)
-    maxiter_int = int(maxiter[()])
-    if not maxiter == maxiter_int or maxiter < 0:
-        raise ValueError(message)
-
-    if not np.all((min <= a) & (a < b) & (b <= max)):
-        raise ValueError('`min <= a < b <= max` must be True (elementwise).')
-
-    return func, a, b, min, max, factor, args, maxiter
-
-
-def _bracket_root(func, a, b=None, *, min=None, max=None, factor=None,
-                  args=(), maxiter=1000):
-    """Bracket the root of a monotonic scalar function of one variable
-
-    This function works elementwise when `a`, `b`, `min`, `max`, `factor`, and
-    the elements of `args` are broadcastable arrays.
-
-    Parameters
-    ----------
-    func : callable
-        The function for which the root is to be bracketed.
-        The signature must be::
-
-            func(x: ndarray, *args) -> ndarray
-
-        where each element of ``x`` is a finite real and ``args`` is a tuple,
-        which may contain an arbitrary number of arrays that are broadcastable
-        with `x`. ``func`` must be an elementwise function: each element
-        ``func(x)[i]`` must equal ``func(x[i])`` for all indices ``i``.
-    a, b : float array_like
-        Starting guess of bracket, which need not contain a root. If `b` is
-        not provided, ``b = a + 1``. Must be broadcastable with one another.
-    min, max : float array_like, optional
-        Minimum and maximum allowable endpoints of the bracket, inclusive. Must
-        be broadcastable with `a` and `b`.
-    factor : float array_like, default: 2
-        The factor used to grow the bracket. See notes for details.
-    args : tuple, optional
-        Additional positional arguments to be passed to `func`.  Must be arrays
-        broadcastable with `a`, `b`, `min`, and `max`. If the callable to be
-        bracketed requires arguments that are not broadcastable with these
-        arrays, wrap that callable with `func` such that `func` accepts
-        only `x` and broadcastable arrays.
-    maxiter : int, optional
-        The maximum number of iterations of the algorithm to perform.
-
-    Returns
-    -------
-    res : OptimizeResult
-        An instance of `scipy.optimize.OptimizeResult` with the following
-        attributes. The descriptions are written as though the values will be
-        scalars; however, if `func` returns an array, the outputs will be
-        arrays of the same shape.
-
-        xl, xr : float
-            The lower and upper ends of the bracket, if the algorithm
-            terminated successfully.
-        fl, fr : float
-            The function value at the lower and upper ends of the bracket.
-        nfev : int
-            The number of function evaluations required to find the bracket.
-            This is distinct from the number of times `func` is *called*
-            because the function may evaluated at multiple points in a single
-            call.
-        nit : int
-            The number of iterations of the algorithm that were performed.
-        status : int
-            An integer representing the exit status of the algorithm.
-
-            - ``0`` : The algorithm produced a valid bracket.
-            - ``-1`` : The bracket expanded to the allowable limits without finding a bracket.
-            - ``-2`` : The maximum number of iterations was reached.
-            - ``-3`` : A non-finite value was encountered.
-            - ``-4`` : Iteration was terminated by `callback`.
-            - ``1`` : The algorithm is proceeding normally (in `callback` only).
-            - ``2`` : A bracket was found in the opposite search direction (in `callback` only).
-
-        success : bool
-            ``True`` when the algorithm terminated successfully (status ``0``).
-
-    Notes
-    -----
-    This function generalizes an algorithm found in pieces throughout
-    `scipy.stats`. The strategy is to iteratively grow the bracket `(l, r)`
-     until ``func(l) < 0 < func(r)``. The bracket grows to the left as follows.
-
-    - If `min` is not provided, the distance between `b` and `l` is iteratively
-      increased by `factor`.
-    - If `min` is provided, the distance between `min` and `l` is iteratively
-      decreased by `factor`. Note that this also *increases* the bracket size.
-
-    Growth of the bracket to the right is analogous.
-
-    Growth of the bracket in one direction stops when the endpoint is no longer
-    finite, the function value at the endpoint is no longer finite, or the
-    endpoint reaches its limiting value (`min` or `max`). Iteration terminates
-    when the bracket stops growing in both directions, the bracket surrounds
-    the root, or a root is found (accidentally).
-
-    If two brackets are found - that is, a bracket is found on both sides in
-    the same iteration, the smaller of the two is returned.
-    If roots of the function are found, both `l` and `r` are set to the
-    leftmost root.
-
-    """  # noqa: E501
-    # Todo:
-    # - find bracket with sign change in specified direction
-    # - Add tolerance
-    # - allow factor < 1?
-
-    callback = None  # works; I just don't want to test it
-    temp = _bracket_root_iv(func, a, b, min, max, factor, args, maxiter)
-    func, a, b, min, max, factor, args, maxiter = temp
-
-    xs = (a, b)
-    temp = _scalar_optimization_initialize(func, xs, args)
-    xs, fs, args, shape, dtype = temp  # line split for PEP8
-
-    # The approach is to treat the left and right searches as though they were
-    # (almost) totally independent one-sided bracket searches. (The interaction
-    # is considered when checking for termination and preparing the result
-    # object.)
-    # `x` is the "moving" end of the bracket
-    x = np.concatenate(xs)
-    f = np.concatenate(fs)
-    n = len(x) // 2
-
-    # `x_last` is the previous location of the moving end of the bracket. If
-    # the signs of `f` and `f_last` are different, `x` and `x_last` form a
-    # bracket.
-    x_last = np.concatenate((x[n:], x[:n]))
-    f_last = np.concatenate((f[n:], f[:n]))
-    # `x0` is the "fixed" end of the bracket.
-    x0 = x_last
-    # We don't need to retain the corresponding function value, since the
-    # fixed end of the bracket is only needed to compute the new value of the
-    # moving end; it is never returned.
-
-    min = np.broadcast_to(min, shape).astype(dtype, copy=False).ravel()
-    max = np.broadcast_to(max, shape).astype(dtype, copy=False).ravel()
-    limit = np.concatenate((min, max))
-
-    factor = np.broadcast_to(factor, shape).astype(dtype, copy=False).ravel()
-    factor = np.concatenate((factor, factor))
-
-    active = np.arange(2*n)
-    args = [np.concatenate((arg, arg)) for arg in args]
-
-    # This is needed due to inner workings of `_scalar_optimization_loop`.
-    # We're abusing it a tiny bit.
-    shape = shape + (2,)
-
-    # `d` is for "distance".
-    # For searches without a limit, the distance between the fixed end of the
-    # bracket `x0` and the moving end `x` will grow by `factor` each iteration.
-    # For searches with a limit, the distance between the `limit` and moving
-    # end of the bracket `x` will shrink by `factor` each iteration.
-    i = np.isinf(limit)
-    ni = ~i
-    d = np.zeros_like(x)
-    d[i] = x[i] - x0[i]
-    d[ni] = limit[ni] - x[ni]
-
-    status = np.full_like(x, _EINPROGRESS, dtype=int)  # in progress
-    nit, nfev = 0, 1  # one function evaluation per side performed above
-
-    work = OptimizeResult(x=x, x0=x0, f=f, limit=limit, factor=factor,
-                          active=active, d=d, x_last=x_last, f_last=f_last,
-                          nit=nit, nfev=nfev, status=status, args=args,
-                          xl=None, xr=None, fl=None, fr=None, n=n)
-    res_work_pairs = [('status', 'status'), ('xl', 'xl'), ('xr', 'xr'),
-                      ('nit', 'nit'), ('nfev', 'nfev'), ('fl', 'fl'),
-                      ('fr', 'fr'), ('x', 'x'), ('f', 'f'),
-                      ('x_last', 'x_last'), ('f_last', 'f_last')]
-
-    def pre_func_eval(work):
-        # Initialize moving end of bracket
-        x = np.zeros_like(work.x)
-
-        # Unlimited brackets grow by `factor` by increasing distance from fixed
-        # end to moving end.
-        i = np.isinf(work.limit)  # indices of unlimited brackets
-        work.d[i] *= work.factor[i]
-        x[i] = work.x0[i] + work.d[i]
-
-        # Limited brackets grow by decreasing the distance from the limit to
-        # the moving end.
-        ni = ~i  # indices of limited brackets
-        work.d[ni] /= work.factor[ni]
-        x[ni] = work.limit[ni] - work.d[ni]
-
-        return x
-
-    def post_func_eval(x, f, work):
-        # Keep track of the previous location of the moving end so that we can
-        # return a narrower bracket. (The alternative is to remember the
-        # original fixed end, but then the bracket would be wider than needed.)
-        work.x_last = work.x
-        work.f_last = work.f
-        work.x = x
-        work.f = f
-
-    def check_termination(work):
-        stop = np.zeros_like(work.x, dtype=bool)
-
-        # Condition 1: a valid bracket (or the root itself) has been found
-        sf = np.sign(work.f)
-        sf_last = np.sign(work.f_last)
-        i = (sf_last == -sf) | (sf_last == 0) | (sf == 0)
-        work.status[i] = _ECONVERGED
-        stop[i] = True
-
-        # Condition 2: the other side's search found a valid bracket.
-        # (If we just found a bracket with the rightward search, we can stop
-        #  the leftward search, and vice-versa.)
-        # To do this, we need to set the status of the other side's search;
-        # this is tricky because `work.status` contains only the *active*
-        # elements, so we don't immediately know the index of the element we
-        # need to set - or even if it's still there. (That search may have
-        # terminated already, e.g. by reaching its `limit`.)
-        # To facilitate this, `work.active` contains a unit integer index of
-        # each search. Index `k` (`k < n)` and `k + n` correspond with a
-        # leftward and rightward search, respectively. Elements are removed
-        # from `work.active` just as they are removed from `work.status`, so
-        # we use `work.active` to help find the right location in
-        # `work.status`.
-        # Get the integer indices of the elements that can also stop
-        also_stop = (work.active[i] + work.n) % (2*work.n)
-        # Check whether they are still active.
-        # To start, we need to find out where in `work.active` they would
-        # appear if they are indeed there.
-        j = np.searchsorted(work.active, also_stop)
-        # If the location exceeds the length of the `work.active`, they are
-        # not there.
-        j = j[j < len(work.active)]
-        # Check whether they are still there.
-        j = j[also_stop == work.active[j]]
-        # Now convert these to boolean indices to use with `work.status`.
-        i = np.zeros_like(stop)
-        i[j] = True  # boolean indices of elements that can also stop
-        i = i & ~stop
-        work.status[i] = _ESTOPONESIDE
-        stop[i] = True
-
-        # Condition 3: moving end of bracket reaches limit
-        i = (work.x == work.limit) & ~stop
-        work.status[i] = _ELIMITS
-        stop[i] = True
-
-        # Condition 4: non-finite value encountered
-        i = ~(np.isfinite(work.x) & np.isfinite(work.f)) & ~stop
-        work.status[i] = _EVALUEERR
-        stop[i] = True
-
-        return stop
-
-    def post_termination_check(work):
-        pass
-
-    def customize_result(res, shape):
-        n = len(res['x']) // 2
-
-        # Because we treat the two one-sided searches as though they were
-        # independent, what we keep track of in `work` and what we want to
-        # return in `res` look quite different. Combine the results from the
-        # two one-sided searches before reporting the results to the user.
-        # - "a" refers to the leftward search (the moving end started at `a`)
-        # - "b" refers to the rightward search (the moving end started at `b`)
-        # - "l" refers to the left end of the bracket (closer to -oo)
-        # - "r" refers to the right end of the bracket (closer to +oo)
-        xal = res['x'][:n]
-        xar = res['x_last'][:n]
-        xbl = res['x_last'][n:]
-        xbr = res['x'][n:]
-
-        fal = res['f'][:n]
-        far = res['f_last'][:n]
-        fbl = res['f_last'][n:]
-        fbr = res['f'][n:]
-
-        # Initialize the brackets and corresponding function values to return
-        # to the user. Brackets may not be valid (e.g. there is no root,
-        # there weren't enough iterations, NaN encountered), but we still need
-        # to return something. One option would be all NaNs, but what I've
-        # chosen here is the left- and right-most points at which the function
-        # has been evaluated. This gives the user some information about what
-        # interval of the real line has been searched and shows that there is
-        # no sign change between the two ends.
-        xl = xal.copy()
-        fl = fal.copy()
-        xr = xbr.copy()
-        fr = fbr.copy()
-
-        # `status` indicates whether the bracket is valid or not. If so,
-        # we want to adjust the bracket we return to be the narrowest possible
-        # given the points at which we evaluated the function.
-        # For example if bracket "a" is valid and smaller than bracket "b" OR
-        # if bracket "a" is valid and bracket "b" is not valid, we want to
-        # return bracket "a" (and vice versa).
-        sa = res['status'][:n]
-        sb = res['status'][n:]
-
-        da = xar - xal
-        db = xbr - xbl
-
-        i1 = ((da <= db) & (sa == 0)) | ((sa == 0) & (sb != 0))
-        i2 = ((db <= da) & (sb == 0)) | ((sb == 0) & (sa != 0))
-
-        xr[i1] = xar[i1]
-        fr[i1] = far[i1]
-        xl[i2] = xbl[i2]
-        fl[i2] = fbl[i2]
-
-        # Finish assembling the result object
-        res['xl'] = xl
-        res['xr'] = xr
-        res['fl'] = fl
-        res['fr'] = fr
-
-        res['nit'] = np.maximum(res['nit'][:n], res['nit'][n:])
-        res['nfev'] = res['nfev'][:n] + res['nfev'][n:]
-        # If the status on one side is zero, the status is zero. In any case,
-        # report the status from one side only.
-        res['status'] = np.choose(sa == 0, (sb, sa))
-        res['success'] = (res['status'] == 0)
-
-        del res['x']
-        del res['f']
-        del res['x_last']
-        del res['f_last']
-
-        return shape[:-1]
-
-    return _scalar_optimization_loop(work, callback, shape,
-                                     maxiter, func, args, dtype,
-                                     pre_func_eval, post_func_eval,
-                                     check_termination, post_termination_check,
-                                     customize_result, res_work_pairs)
-
-
-def _chandrupatla(func, a, b, *, args=(), xatol=_xtol, xrtol=_rtol,
-                  fatol=None, frtol=0, maxiter=_iter, callback=None):
-    """Find the root of an elementwise function using Chandrupatla's algorithm.
-
-    For each element of the output of `func`, `chandrupatla` seeks the scalar
-    root that makes the element 0. This function allows for `a`, `b`, and the
-    output of `func` to be of any broadcastable shapes.
-
-    Parameters
-    ----------
-    func : callable
-        The function whose root is desired. The signature must be::
-
-            func(x: ndarray, *args) -> ndarray
-
-         where each element of ``x`` is a finite real and ``args`` is a tuple,
-         which may contain an arbitrary number of components of any type(s).
-         ``func`` must be an elementwise function: each element ``func(x)[i]``
-         must equal ``func(x[i])`` for all indices ``i``. `_chandrupatla`
-         seeks an array ``x`` such that ``func(x)`` is an array of zeros.
-    a, b : array_like
-        The lower and upper bounds of the root of the function. Must be
-        broadcastable with one another.
-    args : tuple, optional
-        Additional positional arguments to be passed to `func`.
-    xatol, xrtol, fatol, frtol : float, optional
-        Absolute and relative tolerances on the root and function value.
-        See Notes for details.
-    maxiter : int, optional
-        The maximum number of iterations of the algorithm to perform.
-    callback : callable, optional
-        An optional user-supplied function to be called before the first
-        iteration and after each iteration.
-        Called as ``callback(res)``, where ``res`` is an ``OptimizeResult``
-        similar to that returned by `_chandrupatla` (but containing the current
-        iterate's values of all variables). If `callback` raises a
-        ``StopIteration``, the algorithm will terminate immediately and
-        `_chandrupatla` will return a result.
-
-    Returns
-    -------
-    res : OptimizeResult
-        An instance of `scipy.optimize.OptimizeResult` with the following
-        attributes. The descriptions are written as though the values will be
-        scalars; however, if `func` returns an array, the outputs will be
-        arrays of the same shape.
-
-        x : float
-            The root of the function, if the algorithm terminated successfully.
-        nfev : int
-            The number of times the function was called to find the root.
-        nit : int
-            The number of iterations of Chandrupatla's algorithm performed.
-        status : int
-            An integer representing the exit status of the algorithm.
-            ``0`` : The algorithm converged to the specified tolerances.
-            ``-1`` : The algorithm encountered an invalid bracket.
-            ``-2`` : The maximum number of iterations was reached.
-            ``-3`` : A non-finite value was encountered.
-            ``-4`` : Iteration was terminated by `callback`.
-            ``1`` : The algorithm is proceeding normally (in `callback` only).
-        success : bool
-            ``True`` when the algorithm terminated successfully (status ``0``).
-        fun : float
-            The value of `func` evaluated at `x`.
-        xl, xr : float
-            The lower and upper ends of the bracket.
-        fl, fr : float
-            The function value at the lower and upper ends of the bracket.
-
-    Notes
-    -----
-    Implemented based on Chandrupatla's original paper [1]_.
-
-    If ``xl`` and ``xr`` are the left and right ends of the bracket,
-    ``xmin = xl if abs(func(xl)) <= abs(func(xr)) else xr``,
-    and ``fmin0 = min(func(a), func(b))``, then the algorithm is considered to
-    have converged when ``abs(xr - xl) < xatol + abs(xmin) * xrtol`` or
-    ``fun(xmin) <= fatol + abs(fmin0) * frtol``. This is equivalent to the
-    termination condition described in [1]_ with ``xrtol = 4e-10``,
-    ``xatol = 1e-5``, and ``fatol = frtol = 0``. The default values are
-    ``xatol = 2e-12``, ``xrtol = 4 * np.finfo(float).eps``, ``frtol = 0``,
-    and ``fatol`` is the smallest normal number of the ``dtype`` returned
-    by ``func``.
-
-    References
-    ----------
-
-    .. [1] Chandrupatla, Tirupathi R.
-        "A new hybrid quadratic/bisection algorithm for finding the zero of a
-        nonlinear function without using derivatives".
-        Advances in Engineering Software, 28(3), 145-149.
-        https://doi.org/10.1016/s0965-9978(96)00051-8
-
-    See Also
-    --------
-    brentq, brenth, ridder, bisect, newton
-
-    Examples
-    --------
-    >>> from scipy import optimize
-    >>> def f(x, c):
-    ...     return x**3 - 2*x - c
-    >>> c = 5
-    >>> res = optimize._zeros_py._chandrupatla(f, 0, 3, args=(c,))
-    >>> res.x
-    2.0945514818937463
-
-    >>> c = [3, 4, 5]
-    >>> res = optimize._zeros_py._chandrupatla(f, 0, 3, args=(c,))
-    >>> res.x
-    array([1.8932892 , 2.        , 2.09455148])
-
-    """
-    res = _chandrupatla_iv(func, args, xatol, xrtol,
-                           fatol, frtol, maxiter, callback)
-    func, args, xatol, xrtol, fatol, frtol, maxiter, callback = res
-
-    # Initialization
-    xs, fs, args, shape, dtype = _scalar_optimization_initialize(func, (a, b),
-                                                                 args)
-    x1, x2 = xs
-    f1, f2 = fs
-    status = np.full_like(x1, _EINPROGRESS, dtype=int)  # in progress
-    nit, nfev = 0, 2  # two function evaluations performed above
-    xatol = _xtol if xatol is None else xatol
-    xrtol = _rtol if xrtol is None else xrtol
-    fatol = np.finfo(dtype).tiny if fatol is None else fatol
-    frtol = frtol * np.minimum(np.abs(f1), np.abs(f2))
-    work = OptimizeResult(x1=x1, f1=f1, x2=x2, f2=f2, x3=None, f3=None, t=0.5,
-                          xatol=xatol, xrtol=xrtol, fatol=fatol, frtol=frtol,
-                          nit=nit, nfev=nfev, status=status)
-    res_work_pairs = [('status', 'status'), ('x', 'xmin'), ('fun', 'fmin'),
-                      ('nit', 'nit'), ('nfev', 'nfev'), ('xl', 'x1'),
-                      ('fl', 'f1'), ('xr', 'x2'), ('fr', 'f2')]
-
-    def pre_func_eval(work):
-        # [1] Figure 1 (first box)
-        x = work.x1 + work.t * (work.x2 - work.x1)
-        return x
-
-    def post_func_eval(x, f, work):
-        # [1] Figure 1 (first diamond and boxes)
-        # Note: y/n are reversed in figure; compare to BASIC in appendix
-        work.x3, work.f3 = work.x2.copy(), work.f2.copy()
-        j = np.sign(f) == np.sign(work.f1)
-        nj = ~j
-        work.x3[j], work.f3[j] = work.x1[j], work.f1[j]
-        work.x2[nj], work.f2[nj] = work.x1[nj], work.f1[nj]
-        work.x1, work.f1 = x, f
-
-    def check_termination(work):
-        # [1] Figure 1 (second diamond)
-        # Check for all terminal conditions and record statuses.
-
-        # See [1] Section 4 (first two sentences)
-        i = np.abs(work.f1) < np.abs(work.f2)
-        work.xmin = np.choose(i, (work.x2, work.x1))
-        work.fmin = np.choose(i, (work.f2, work.f1))
-        stop = np.zeros_like(work.x1, dtype=bool)  # termination condition met
-
-        # This is the convergence criterion used in bisect. Chandrupatla's
-        # criterion is equivalent to this except with a factor of 4 on `xrtol`.
-        work.dx = abs(work.x2 - work.x1)
-        work.tol = abs(work.xmin) * work.xrtol + work.xatol
-        i = work.dx < work.tol
-        # Modify in place to incorporate tolerance on function value. Note that
-        # `frtol` has been redefined as `frtol = frtol * np.minimum(f1, f2)`,
-        # where `f1` and `f2` are the function evaluated at the original ends of
-        # the bracket.
-        i |= np.abs(work.fmin) <= work.fatol + work.frtol
-        work.status[i] = _ECONVERGED
-        stop[i] = True
-
-        i = (np.sign(work.f1) == np.sign(work.f2)) & ~stop
-        work.xmin[i], work.fmin[i], work.status[i] = np.nan, np.nan, _ESIGNERR
-        stop[i] = True
-
-        i = ~((np.isfinite(work.x1) & np.isfinite(work.x2)
-               & np.isfinite(work.f1) & np.isfinite(work.f2)) | stop)
-        work.xmin[i], work.fmin[i], work.status[i] = np.nan, np.nan, _EVALUEERR
-        stop[i] = True
-
-        return stop
-
-    def post_termination_check(work):
-        # [1] Figure 1 (third diamond and boxes / Equation 1)
-        xi1 = (work.x1 - work.x2) / (work.x3 - work.x2)
-        phi1 = (work.f1 - work.f2) / (work.f3 - work.f2)
-        alpha = (work.x3 - work.x1) / (work.x2 - work.x1)
-        j = ((1 - np.sqrt(1 - xi1)) < phi1) & (phi1 < np.sqrt(xi1))
-
-        f1j, f2j, f3j, alphaj = work.f1[j], work.f2[j], work.f3[j], alpha[j]
-        t = np.full_like(alpha, 0.5)
-        t[j] = (f1j / (f1j - f2j) * f3j / (f3j - f2j)
-                - alphaj * f1j / (f3j - f1j) * f2j / (f2j - f3j))
-
-        # [1] Figure 1 (last box; see also BASIC in appendix with comment
-        # "Adjust T Away from the Interval Boundary")
-        tl = 0.5 * work.tol / work.dx
-        work.t = np.clip(t, tl, 1 - tl)
-
-    def customize_result(res, shape):
-        xl, xr, fl, fr = res['xl'], res['xr'], res['fl'], res['fr']
-        i = res['xl'] < res['xr']
-        res['xl'] = np.choose(i, (xr, xl))
-        res['xr'] = np.choose(i, (xl, xr))
-        res['fl'] = np.choose(i, (fr, fl))
-        res['fr'] = np.choose(i, (fl, fr))
-        return shape
-
-    return _scalar_optimization_loop(work, callback, shape,
-                                     maxiter, func, args, dtype,
-                                     pre_func_eval, post_func_eval,
-                                     check_termination, post_termination_check,
-                                     customize_result, res_work_pairs)
-
-
-def _chandrupatla_iv(func, args, xatol, xrtol,
-                     fatol, frtol, maxiter, callback):
-    # Input validation for `_chandrupatla`
-
-    if not callable(func):
-        raise ValueError('`func` must be callable.')
-
-    if not np.iterable(args):
-        args = (args,)
-
-    tols = np.asarray([xatol if xatol is not None else 1,
-                       xrtol if xrtol is not None else 1,
-                       fatol if fatol is not None else 1,
-                       frtol if frtol is not None else 1])
-    if (not np.issubdtype(tols.dtype, np.number) or np.any(tols < 0)
-            or np.any(np.isnan(tols)) or tols.shape != (4,)):
-        raise ValueError('Tolerances must be non-negative scalars.')
-
-    maxiter_int = int(maxiter)
-    if maxiter != maxiter_int or maxiter < 0:
-        raise ValueError('`maxiter` must be a non-negative integer.')
-
-    if callback is not None and not callable(callback):
-        raise ValueError('`callback` must be callable.')
-
-    return func, args, xatol, xrtol, fatol, frtol, maxiter, callback
-
-
-def _differentiate_iv(func, x, args, atol, rtol, maxiter, order,
-                      initial_step, step_factor, step_direction, callback):
-    # Input validation for `_differentiate`
-
-    if not callable(func):
-        raise ValueError('`func` must be callable.')
-
-    # x has more complex IV that is taken care of during initialization
-    x = np.asarray(x)
-    dtype = x.dtype if np.issubdtype(x.dtype, np.inexact) else np.float64
-
-    if not np.iterable(args):
-        args = (args,)
-
-    if atol is None:
-        atol = np.finfo(dtype).tiny
-
-    if rtol is None:
-        rtol = np.sqrt(np.finfo(dtype).eps)
-
-    message = 'Tolerances and step parameters must be non-negative scalars.'
-    tols = np.asarray([atol, rtol, initial_step, step_factor])
-    if (not np.issubdtype(tols.dtype, np.number)
-            or np.any(tols < 0)
-            or tols.shape != (4,)):
-        raise ValueError(message)
-    initial_step, step_factor = tols[2:].astype(dtype)
-
-    maxiter_int = int(maxiter)
-    if maxiter != maxiter_int or maxiter <= 0:
-        raise ValueError('`maxiter` must be a positive integer.')
-
-    order_int = int(order)
-    if order_int != order or order <= 0:
-        raise ValueError('`order` must be a positive integer.')
-
-    step_direction = np.sign(step_direction).astype(dtype)
-    x, step_direction = np.broadcast_arrays(x, step_direction)
-    x, step_direction = x[()], step_direction[()]
-
-    if callback is not None and not callable(callback):
-        raise ValueError('`callback` must be callable.')
-
-    return (func, x, args, atol, rtol, maxiter_int, order_int, initial_step,
-            step_factor, step_direction, callback)
-
-
-def _differentiate(func, x, *, args=(), atol=None, rtol=None, maxiter=10,
-                   order=8, initial_step=0.5, step_factor=2.0,
-                   step_direction=0, callback=None):
-    """Evaluate the derivative of an elementwise scalar function numerically.
-
-    Parameters
-    ----------
-    func : callable
-        The function whose derivative is desired. The signature must be::
-
-            func(x: ndarray, *args) -> ndarray
-
-         where each element of ``x`` is a finite real and ``args`` is a tuple,
-         which may contain an arbitrary number of arrays that are broadcastable
-         with `x`. ``func`` must be an elementwise function: each element
-         ``func(x)[i]`` must equal ``func(x[i])`` for all indices ``i``.
-    x : array_like
-        Abscissae at which to evaluate the derivative.
-    args : tuple, optional
-        Additional positional arguments to be passed to `func`. Must be arrays
-        broadcastable with `x`. If the callable to be differentiated requires
-        arguments that are not broadcastable with `x`, wrap that callable with
-        `func`. See Examples.
-    atol, rtol : float, optional
-        Absolute and relative tolerances for the stopping condition: iteration
-        will stop when ``res.error < atol + rtol * abs(res.df)``. The default
-        `atol` is the smallest normal number of the appropriate dtype, and
-        the default `rtol` is the square root of the precision of the
-        appropriate dtype.
-    order : int, default: 8
-        The (positive integer) order of the finite difference formula to be
-        used. Odd integers will be rounded up to the next even integer.
-    initial_step : float, default: 0.5
-        The (absolute) initial step size for the finite difference derivative
-        approximation.
-    step_factor : float, default: 2.0
-        The factor by which the step size is *reduced* in each iteration; i.e.
-        the step size in iteration 1 is ``initial_step/step_factor``. If
-        ``step_factor < 1``, subsequent steps will be greater than the initial
-        step; this may be useful if steps smaller than some threshold are
-        undesirable (e.g. due to subtractive cancellation error).
-    maxiter : int, default: 10
-        The maximum number of iterations of the algorithm to perform. See
-        notes.
-    step_direction : array_like
-        An array representing the direction of the finite difference steps (for
-        use when `x` lies near to the boundary of the domain of the function.)
-        Must be broadcastable with `x` and all `args`.
-        Where 0 (default), central differences are used; where negative (e.g.
-        -1), steps are non-positive; and where positive (e.g. 1), all steps are
-        non-negative.
-    callback : callable, optional
-        An optional user-supplied function to be called before the first
-        iteration and after each iteration.
-        Called as ``callback(res)``, where ``res`` is an ``OptimizeResult``
-        similar to that returned by `_differentiate` (but containing the
-        current iterate's values of all variables). If `callback` raises a
-        ``StopIteration``, the algorithm will terminate immediately and
-        `_differentiate` will return a result.
-
-    Returns
-    -------
-    res : OptimizeResult
-        An instance of `scipy.optimize.OptimizeResult` with the following
-        attributes. (The descriptions are written as though the values will be
-        scalars; however, if `func` returns an array, the outputs will be
-        arrays of the same shape.)
-
-        success : bool
-            ``True`` when the algorithm terminated successfully (status ``0``).
-        status : int
-            An integer representing the exit status of the algorithm.
-            ``0`` : The algorithm converged to the specified tolerances.
-            ``-1`` : The error estimate increased, so iteration was terminated.
-            ``-2`` : The maximum number of iterations was reached.
-            ``-3`` : A non-finite value was encountered.
-            ``-4`` : Iteration was terminated by `callback`.
-            ``1`` : The algorithm is proceeding normally (in `callback` only).
-        df : float
-            The derivative of `func` at `x`, if the algorithm terminated
-            successfully.
-        error : float
-            An estimate of the error: the magnitude of the difference between
-            the current estimate of the derivative and the estimate in the
-            previous iteration.
-        nit : int
-            The number of iterations performed.
-        nfev : int
-            The number of points at which `func` was evaluated.
-        x : float
-            The value at which the derivative of `func` was evaluated
-            (after broadcasting with `args` and `step_direction`).
-
-    Notes
-    -----
-    The implementation was inspired by jacobi [1]_, numdifftools [2]_, and
-    DERIVEST [3]_, but the implementation follows the theory of Taylor series
-    more straightforwardly (and arguably naively so).
-    In the first iteration, the derivative is estimated using a finite
-    difference formula of order `order` with maximum step size `initial_step`.
-    Each subsequent iteration, the maximum step size is reduced by
-    `step_factor`, and the derivative is estimated again until a termination
-    condition is reached. The error estimate is the magnitude of the difference
-    between the current derivative approximation and that of the previous
-    iteration.
-
-    The stencils of the finite difference formulae are designed such that
-    abscissae are "nested": after `func` is evaluated at ``order + 1``
-    points in the first iteration, `func` is evaluated at only two new points
-    in each subsequent iteration; ``order - 1`` previously evaluated function
-    values required by the finite difference formula are reused, and two
-    function values (evaluations at the points furthest from `x`) are unused.
-
-    Step sizes are absolute. When the step size is small relative to the
-    magnitude of `x`, precision is lost; for example, if `x` is ``1e20``, the
-    default initial step size of ``0.5`` cannot be resolved. Accordingly,
-    consider using larger initial step sizes for large magnitudes of `x`.
-
-    The default tolerances are challenging to satisfy at points where the
-    true derivative is exactly zero. If the derivative may be exactly zero,
-    consider specifying an absolute tolerance (e.g. ``atol=1e-16``) to
-    improve convergence.
-
-    References
-    ----------
-    [1]_ Hans Dembinski (@HDembinski). jacobi.
-         https://github.com/HDembinski/jacobi
-    [2]_ Per A. Brodtkorb and John D'Errico. numdifftools.
-         https://numdifftools.readthedocs.io/en/latest/
-    [3]_ John D'Errico. DERIVEST: Adaptive Robust Numerical Differentiation.
-         https://www.mathworks.com/matlabcentral/fileexchange/13490-adaptive-robust-numerical-differentiation
-    [4]_ Numerical Differentition. Wikipedia.
-         https://en.wikipedia.org/wiki/Numerical_differentiation
-
-    Examples
-    --------
-    Evaluate the derivative of ``np.exp`` at several points ``x``.
-
-    >>> import numpy as np
-    >>> from scipy.optimize._zeros_py import _differentiate
-    >>> f = np.exp
-    >>> df = np.exp  # true derivative
-    >>> x = np.linspace(1, 2, 5)
-    >>> res = _differentiate(f, x)
-    >>> res.df  # approximation of the derivative
-    array([2.71828183, 3.49034296, 4.48168907, 5.75460268, 7.3890561 ])
-    >>> res.error  # estimate of the error
-    array(
-        [7.12940817e-12, 9.16688947e-12, 1.17594823e-11, 1.50972568e-11, 1.93942640e-11]
-    )
-    >>> abs(res.df - df(x))  # true error
-    array(
-        [3.06421555e-14, 3.01980663e-14, 5.06261699e-14, 6.30606678e-14, 8.34887715e-14]
-    )
-
-    Show the convergence of the approximation as the step size is reduced.
-    Each iteration, the step size is reduced by `step_factor`, so for
-    sufficiently small initial step, each iteration reduces the error by a
-    factor of ``1/step_factor**order`` until finite precision arithmetic
-    inhibits further improvement.
-
-    >>> iter = list(range(1, 12))  # maximum iterations
-    >>> hfac = 2  # step size reduction per iteration
-    >>> hdir = [-1, 0, 1]  # compare left-, central-, and right- steps
-    >>> order = 4  # order of differentiation formula
-    >>> x = 1
-    >>> ref = df(x)
-    >>> errors = []  # true error
-    >>> for i in iter:
-    ...     res = _differentiate(f, x, maxiter=i, step_factor=hfac,
-    ...                          step_direction=hdir, order=order,
-    ...                          atol=0, rtol=0)  # prevent early termination
-    ...     errors.append(abs(res.df - ref))
-    >>> errors = np.array(errors)
-    >>> plt.semilogy(iter, errors[:, 0], label='left differences')
-    >>> plt.semilogy(iter, errors[:, 1], label='central differences')
-    >>> plt.semilogy(iter, errors[:, 2], label='right differences')
-    >>> plt.xlabel('iteration')
-    >>> plt.ylabel('error')
-    >>> plt.legend()
-    >>> plt.show()
-    >>> (errors[1, 1] / errors[0, 1], 1 / hfac**order)
-    (0.06215223140159822, 0.0625)
-
-    The implementation is vectorized over `x`, `step_direction`, and `args`.
-    The function is evaluated once before the first iteration to perform input
-    validation and standardization, and once per iteration thereafter.
-
-    >>> def f(x, p):
-    ...     print('here')
-    ...     f.nit += 1
-    ...     return x**p
-    >>> f.nit = 0
-    >>> def df(x, p):
-    ...     return p*x**(p-1)
-    >>> x = np.arange(1, 5)
-    >>> p = np.arange(1, 6).reshape((-1, 1))
-    >>> hdir = np.arange(-1, 2).reshape((-1, 1, 1))
-    >>> res = _differentiate(f, x, args=(p,), step_direction=hdir, maxiter=1)
-    >>> np.allclose(res.df, df(x, p))
-    True
-    >>> res.df.shape
-    (3, 5, 4)
-    >>> f.nit
-    2
-
-    """
-    # TODO (followup):
-    #  - investigate behavior at saddle points
-    #  - array initial_step / step_factor?
-    #  - multivariate functions?
-    #  - vector-valued functions?
-
-    res = _differentiate_iv(func, x, args, atol, rtol, maxiter, order,
-                            initial_step, step_factor, step_direction, callback)
-    func, x, args, atol, rtol, maxiter, order, h0, fac, hdir, callback = res
-
-    # Initialization
-    # Since f(x) (no step) is not needed for central differences, it may be
-    # possible to eliminate this function evaluation. However, it's useful for
-    # input validation and standardization, and everything else is designed to
-    # reduce function calls, so let's keep it simple.
-    xs, fs, args, shape, dtype = _scalar_optimization_initialize(func, (x,), args)
-    x, f = xs[0], fs[0]
-    df = np.full_like(f, np.nan)
-    # Ideally we'd broadcast the shape of `hdir` in `_scalar_opt_init`, but
-    # it's simpler to do it here than to generalize `_scalar_opt_init` further.
-    # `hdir` and `x` are already broadcasted in `_differentiate_iv`, so we know
-    # that `hdir` can be broadcasted to the final shape.
-    hdir = np.broadcast_to(hdir, shape).flatten()
-
-    status = np.full_like(x, _EINPROGRESS, dtype=int)  # in progress
-    nit, nfev = 0, 1  # one function evaluations performed above
-    # Boolean indices of left, central, right, and (all) one-sided steps
-    il = hdir < 0
-    ic = hdir == 0
-    ir = hdir > 0
-    io = il | ir
-
-    # Most of these attributes are reasonably obvious, but:
-    # - `fs` holds all the function values of all active `x`. The zeroth
-    #   axis corresponds with active points `x`, the first axis corresponds
-    #   with the different steps (in the order described in
-    #   `_differentiate_weights`).
-    # - `terms` (which could probably use a better name) is half the `order`,
-    #   which is always even.
-    work = OptimizeResult(x=x, df=df, fs=f[:, np.newaxis], error=np.nan, h=h0,
-                          df_last=np.nan, error_last=np.nan, h0=h0, fac=fac,
-                          atol=atol, rtol=rtol, nit=nit, nfev=nfev,
-                          status=status, dtype=dtype, terms=(order+1)//2,
-                          hdir=hdir, il=il, ic=ic, ir=ir, io=io)
-    # This is the correspondence between terms in the `work` object and the
-    # final result. In this case, the mapping is trivial. Note that `success`
-    # is prepended automatically.
-    res_work_pairs = [('status', 'status'), ('df', 'df'), ('error', 'error'),
-                      ('nit', 'nit'), ('nfev', 'nfev'), ('x', 'x')]
-
-    def pre_func_eval(work):
-        """Determine the abscissae at which the function needs to be evaluated.
-
-        See `_differentiate_weights` for a description of the stencil (pattern
-        of the abscissae).
-
-        In the first iteration, there is only one stored function value in
-        `work.fs`, `f(x)`, so we need to evaluate at `order` new points. In
-        subsequent iterations, we evaluate at two new points. Note that
-        `work.x` is always flattened into a 1D array after broadcasting with
-        all `args`, so we add a new axis at the end and evaluate all point
-        in one call to the function.
-
-        For improvement:
-        - Consider measuring the step size actually taken, since `(x + h) - x`
-          is not identically equal to `h` with floating point arithmetic.
-        - Adjust the step size automatically if `x` is too big to resolve the
-          step.
-        - We could probably save some work if there are no central difference
-          steps or no one-sided steps.
-        """
-        n = work.terms  # half the order
-        h = work.h  # step size
-        c = work.fac  # step reduction factor
-        d = c**0.5  # square root of step reduction factor (one-sided stencil)
-        # Note - no need to be careful about dtypes until we allocate `x_eval`
-
-        if work.nit == 0:
-            hc = h / c**np.arange(n)
-            hc = np.concatenate((-hc[::-1], hc))
-        else:
-            hc = np.asarray([-h, h]) / c**(n-1)
-
-        if work.nit == 0:
-            hr = h / d**np.arange(2*n)
-        else:
-            hr = np.asarray([h, h/d]) / c**(n-1)
-
-        n_new = 2*n if work.nit == 0 else 2  # number of new abscissae
-        x_eval = np.zeros((len(work.hdir), n_new), dtype=work.dtype)
-        il, ic, ir = work.il, work.ic, work.ir
-        x_eval[ir] = work.x[ir, np.newaxis] + hr
-        x_eval[ic] = work.x[ic, np.newaxis] + hc
-        x_eval[il] = work.x[il, np.newaxis] - hr
-        return x_eval
-
-    def post_func_eval(x, f, work):
-        """ Estimate the derivative and error from the function evaluations
-
-        As in `pre_func_eval`: in the first iteration, there is only one stored
-        function value in `work.fs`, `f(x)`, so we need to add the `order` new
-        points. In subsequent iterations, we add two new points. The tricky
-        part is getting the order to match that of the weights, which is
-        described in `_differentiate_weights`.
-
-        For improvement:
-        - Change the order of the weights (and steps in `pre_func_eval`) to
-          simplify `work_fc` concatenation and eliminate `fc` concatenation.
-        - It would be simple to do one-step Richardson extrapolation with `df`
-          and `df_last` to increase the order of the estimate and/or improve
-          the error estimate.
-        - Process the function evaluations in a more numerically favorable
-          way. For instance, combining the pairs of central difference evals
-          into a second-order approximation and using Richardson extrapolation
-          to produce a higher order approximation seemed to retain accuracy up
-          to very high order.
-        - Alternatively, we could use `polyfit` like Jacobi. An advantage of
-          fitting polynomial to more points than necessary is improved noise
-          tolerance.
-        """
-        n = work.terms
-        n_new = n if work.nit == 0 else 1
-        il, ic, io = work.il, work.ic, work.io
-
-        # Central difference
-        # `work_fc` is *all* the points at which the function has been evaluated
-        # `fc` is the points we're using *this iteration* to produce the estimate
-        work_fc = (f[ic, :n_new], work.fs[ic, :], f[ic, -n_new:])
-        work_fc = np.concatenate(work_fc, axis=-1)
-        if work.nit == 0:
-            fc = work_fc
-        else:
-            fc = (work_fc[:, :n], work_fc[:, n:n+1], work_fc[:, -n:])
-            fc = np.concatenate(fc, axis=-1)
-
-        # One-sided difference
-        work_fo = np.concatenate((work.fs[io, :], f[io, :]), axis=-1)
-        if work.nit == 0:
-            fo = work_fo
-        else:
-            fo = np.concatenate((work_fo[:, 0:1], work_fo[:, -2*n:]), axis=-1)
-
-        work.fs = np.zeros((len(ic), work.fs.shape[-1] + 2*n_new))
-        work.fs[ic] = work_fc
-        work.fs[io] = work_fo
-
-        wc, wo = _differentiate_weights(work, n)
-        work.df_last = work.df.copy()
-        work.df[ic] = fc @ wc / work.h
-        work.df[io] = fo @ wo / work.h
-        work.df[il] *= -1
-
-        work.h /= work.fac
-        work.error_last = work.error
-        # Simple error estimate - the difference in derivative estimates between
-        # this iteration and the last. This is typically conservative because if
-        # convergence has begin, the true error is much closer to the difference
-        # between the current estimate and the *next* error estimate. However,
-        # we could use Richarson extrapolation to produce an error estimate that
-        # is one order higher, and take the difference between that and
-        # `work.df` (which would just be constant factor that depends on `fac`.)
-        work.error = abs(work.df - work.df_last)
-
-    def check_termination(work):
-        """Terminate due to convergence, non-finite values, or error increase"""
-        stop = np.zeros_like(work.df).astype(bool)
-
-        i = work.error < work.atol + work.rtol*abs(work.df)
-        work.status[i] = _ECONVERGED
-        stop[i] = True
-
-        if work.nit > 0:
-            i = ~((np.isfinite(work.x) & np.isfinite(work.df)) | stop)
-            work.df[i], work.status[i] = np.nan, _EVALUEERR
-            stop[i] = True
-
-        # With infinite precision, there is a step size below which
-        # all smaller step sizes will reduce the error. But in floating point
-        # arithmetic, catastrophic cancellation will begin to cause the error
-        # to increase again. This heuristic tries to avoid step sizes that are
-        # too small. There may be more theoretically sound approaches for
-        # detecting a step size that minimizes the total error, but this
-        # heuristic seems simple and effective.
-        i = (work.error > work.error_last*10) & ~stop
-        work.status[i] = _EERRORINCREASE
-        stop[i] = True
-
-        return stop
-
-    def post_termination_check(work):
-        return
-
-    def customize_result(res, shape):
-        return shape
-
-    return _scalar_optimization_loop(work, callback, shape,
-                                     maxiter, func, args, dtype,
-                                     pre_func_eval, post_func_eval,
-                                     check_termination, post_termination_check,
-                                     customize_result, res_work_pairs)
-
-
-def _differentiate_weights(work, n):
-    # This produces the weights of the finite difference formula for a given
-    # stencil. In experiments, use of a second-order central difference formula
-    # with Richardson extrapolation was more accurate numerically, but it was
-    # more complicated, and it would have become even more complicated when
-    # adding support for one-sided differences. However, now that all the
-    # function evaluation values are stored, they can be processed in whatever
-    # way is desired to produce the derivative estimate. We leave alternative
-    # approaches to future work. To be more self-contained, here is the theory
-    # for deriving the weights below.
-    #
-    # Recall that the Taylor expansion of a univariate, scalar-values function
-    # about a point `x` may be expressed as:
-    #      f(x + h)  =     f(x) + f'(x)*h + f''(x)/2!*h**2  + O(h**3)
-    # Suppose we evaluate f(x), f(x+h), and f(x-h).  We have:
-    #      f(x)      =     f(x)
-    #      f(x + h)  =     f(x) + f'(x)*h + f''(x)/2!*h**2  + O(h**3)
-    #      f(x - h)  =     f(x) - f'(x)*h + f''(x)/2!*h**2  + O(h**3)
-    # We can solve for weights `wi` such that:
-    #   w1*f(x)      = w1*(f(x))
-    # + w2*f(x + h)  = w2*(f(x) + f'(x)*h + f''(x)/2!*h**2) + O(h**3)
-    # + w3*f(x - h)  = w3*(f(x) - f'(x)*h + f''(x)/2!*h**2) + O(h**3)
-    #                =     0    + f'(x)*h + 0               + O(h**3)
-    # Then
-    #     f'(x) ~ (w1*f(x) + w2*f(x+h) + w3*f(x-h))/h
-    # is a finite difference derivative approximation with error O(h**2),
-    # and so it is said to be a "second-order" approximation. Under certain
-    # conditions (e.g. well-behaved function, `h` sufficiently small), the
-    # error in the approximation will decrease with h**2; that is, if `h` is
-    # reduced by a factor of 2, the error is reduced by a factor of 4.
-    #
-    # By default, we use eighth-order formulae. Our central-difference formula
-    # uses abscissae:
-    #   x-h/c**3, x-h/c**2, x-h/c, x-h, x, x+h, x+h/c, x+h/c**2, x+h/c**3
-    # where `c` is the step factor. (Typically, the step factor is greater than
-    # one, so the outermost points - as written above - are actually closest to
-    # `x`.) This "stencil" is chosen so that each iteration, the step can be
-    # reduced by the factor `c`, and most of the function evaluations can be
-    # reused with the new step size. For example, in the next iteration, we
-    # will have:
-    #   x-h/c**4, x-h/c**3, x-h/c**2, x-h/c, x, x+h/c, x+h/c**2, x+h/c**3, x+h/c**4
-    # We do not reuse `x-h` and `x+h` for the new derivative estimate.
-    # While this would increase the order of the formula and thus the
-    # theoretical convergence rate, it is also less stable numerically.
-    # (As noted above, there are other ways of processing the values that are
-    # more stable. Thus, even now we store `f(x-h)` and `f(x+h)` in `work.fs`
-    # to simplify future development of this sort of improvement.)
-    #
-    # The (right) one-sided formula is produced similarly using abscissae
-    #   x, x+h, x+h/d, x+h/d**2, ..., x+h/d**6, x+h/d**7, x+h/d**7
-    # where `d` is the square root of `c`. (The left one-sided formula simply
-    # uses -h.) When the step size is reduced by factor `c = d**2`, we have
-    # abscissae:
-    #   x, x+h/d**2, x+h/d**3..., x+h/d**8, x+h/d**9, x+h/d**9
-    # `d` is chosen as the square root of `c` so that the rate of the step-size
-    # reduction is the same per iteration as in the central difference case.
-    # Note that because the central difference formulas are inherently of even
-    # order, for simplicity, we use only even-order formulas for one-sided
-    # differences, too.
-
-    # It's possible for the user to specify `fac` in, say, double precision but
-    # `x` and `args` in single precision. `fac` gets converted to single
-    # precision, but we should always use double precision for the intermediate
-    # calculations here to avoid additional error in the weights.
-    fac = work.fac.astype(np.float64)
-
-    # Note that if the user switches back to floating point precision with
-    # `x` and `args`, then `fac` will not necessarily equal the (lower
-    # precision) cached `_differentiate_weights.fac`, and the weights will
-    # need to be recalculated. This could be fixed, but it's late, and of
-    # low consequence.
-    if fac != _differentiate_weights.fac:
-        _differentiate_weights.central = []
-        _differentiate_weights.right = []
-        _differentiate_weights.fac = fac
-
-    if len(_differentiate_weights.central) != 2*n + 1:
-        # Central difference weights. Consider refactoring this; it could
-        # probably be more compact.
-        i = np.arange(-n, n + 1)
-        p = np.abs(i) - 1.  # center point has power `p` -1, but sign `s` is 0
-        s = np.sign(i)
-
-        h = s / fac ** p
-        A = np.vander(h, increasing=True).T
-        b = np.zeros(2*n + 1)
-        b[1] = 1
-        weights = np.linalg.solve(A, b)
-
-        # Enforce identities to improve accuracy
-        weights[n] = 0
-        for i in range(n):
-            weights[-i-1] = -weights[i]
-
-        # Cache the weights. We only need to calculate them once unless
-        # the step factor changes.
-        _differentiate_weights.central = weights
-
-        # One-sided difference weights. The left one-sided weights (with
-        # negative steps) are simply the negative of the right one-sided
-        # weights, so no need to compute them separately.
-        i = np.arange(2*n + 1)
-        p = i - 1.
-        s = np.sign(i)
-
-        h = s / np.sqrt(fac) ** p
-        A = np.vander(h, increasing=True).T
-        b = np.zeros(2 * n + 1)
-        b[1] = 1
-        weights = np.linalg.solve(A, b)
-
-        _differentiate_weights.right = weights
-
-    return (_differentiate_weights.central.astype(work.dtype, copy=False),
-            _differentiate_weights.right.astype(work.dtype, copy=False))
-_differentiate_weights.central = []
-_differentiate_weights.right = []
-_differentiate_weights.fac = None
