+++ scipy/optimize/meson.build
@@ -254,11 +254,13 @@ py3.install_sources(
 py3.install_sources([
   '__init__.py',
   '_basinhopping.py',
+  '_bracket.py',
   '_chandrupatla.py',
   '_cobyla_py.py',
   '_constraints.py',
   '_differentiable_functions.py',
   '_differentialevolution.py',
+  '_differentiate.py',
   '_direct_py.py',
   '_dcsrch.py',
   '_dual_annealing.py',
