+++ scipy/optimize/tests/meson.build
@@ -9,12 +9,14 @@ py3.install_sources([
     'test__root.py',
     'test__shgo.py',
     'test__spectral.py',
+    'test_bracket.py',
     'test_chandrupatla.py',
     'test_cobyla.py',
     'test_constraint_conversion.py',
     'test_constraints.py',
     'test_cython_optimize.py',
     'test_differentiable_functions.py',
+    'test_differentiate.py',
     'test_direct.py',
     'test_hessian_update_strategy.py',
     'test_isotonic_regression.py',
